Backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
dj-rest-auth==4.0.0
djangorestframework-simplejwt==5.3.0
Pillow==12.1.0
python-decouple==3.8
stripe==6.5.0
gunicorn==21.2.0
whitenoise==6.11.0